feat: ViewPagerの継承クラスではなくスワイプを検知しイベントを送信するカスタムViewを作成した
