+++ modules/features/media/src/main/java/net/pantasystem/milktea/media/ImageFragment.kt
@@ -56,6 +56,10 @@ class ImageFragment : Fragment(R.layout.fragment_image){
             return
         }
 
+        binding.swipeFinishLayout.setOnFinishEventListener {
+            requireActivity().finish()
+        }
+
         Glide.with(view.context).let {
             if (uri == null) {
                 it.load(url)
