+++ modules/features/media/src/main/res/layout/fragment_image.xml
@@ -1,8 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <com.github.chrisbanes.photoview.PhotoView
+    <net.pantasystem.milktea.media.SwipeFinishLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:id="@+id/swipeFinishLayout">
+        <com.github.chrisbanes.photoview.PhotoView
             android:id="@+id/imageView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"/>
+    </net.pantasystem.milktea.media.SwipeFinishLayout>
+
 </layout>
