+++ modules/features/media/src/main/res/layout/fragment_player.xml
@@ -2,8 +2,13 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:orientation="vertical" android:layout_width="match_parent"
         android:layout_height="match_parent">
-    <com.google.android.exoplayer2.ui.StyledPlayerView
+    <net.pantasystem.milktea.media.SwipeFinishLayout
+        android:id="@+id/swipeFinishLayout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+        <com.google.android.exoplayer2.ui.StyledPlayerView
             android:id="@+id/player_view"
             android:layout_width="match_parent"
             android:layout_height="match_parent"/>
+    </net.pantasystem.milktea.media.SwipeFinishLayout>
 </LinearLayout>
